delete QueryExecutor interface
